requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.47.0
gast==0.3.3
google-auth==2.26.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.0
h5py==2.10.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.5.0
networkx==3.1
numpy==1.18.5
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
protobuf==3.20.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.0
scipy==1.4.1
setuptools-scm==8.0.4
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tifffile==2023.7.10
tomli==2.0.1
typing_extensions==4.5.0
unet @ git+https://github.com/jakeret/unet.git@f557a51b6f95aae6848cab6141e6cae573934bf8
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
